Backport constraint fix for movdf instruction:
Note, this is for the 'movdf_internal1', and not the 'movdf_internal2' constraint, even though the original commit mentions it fixes 'movdf_interal2'. see
https://gcc.gnu.org/git/?p=gcc.git;a=blobdiff;f=gcc/config/mips/mips.md;h=4ada75fde8eb122b576960ed48172731228b2c2a;hp=0f3efb65b7ed2431130ab74e2c2a7074353bd74d;hb=7798dc704cc34e5142ee6288ecf833a899166da9;hpb=d6e0e0532b1f293b61c28b22518b6e7a7efb5a5a
for the original patch, and see https://gcc.gnu.org/bugzilla/show_bug.cgi?format=multiple&id=9092 for the corresponding bug report.

Unfortunately, building cacos() in newlib still fails with an ICE, even with this patch.